+++ inst/run/dev.R
@@ -1,0 +1,9 @@
+leprechaun::build()
+
+pkgload::load_all(
+	path = "../../",
+	reset = TRUE,
+	helpers = FALSE
+)
+
+run()