+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Assets supports added for serving `type = module`
 - Add `tag2` convenience function to htmltools plugin
+- Add `run_dev` function that runs the build step before
+the launching the app.
 
 # leprechaun 1.0.0.9000